🏗️ Move components related to reaction feature into new dedicated folder
