+++ src/features/activity-feed/reaction/components/PostReactionsCounter/style.scss
@@ -1,6 +1,6 @@
-@use '../../styles/settings/breakpoints' as b;
-@use '../../styles/settings/sizes' as s;
-@use '../../styles/settings/colors' as c;
+@use '../../../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/sizes' as s;
+@use '../../../../../styles/settings/colors' as c;
 
 .c-reaction-post{
