+++ src/features/activity-feed/reaction/components/ReactionButton/index.jsx
@@ -4,9 +4,9 @@ import Popover from '@mui/material/Popover'
 import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
 import { useDispatch, useSelector } from 'react-redux'
-import { getPostReactions } from '../../../redux/selectors/feed'
-import { getUserId } from '../../../features/user/store/userSelectors'
-import { createReaction, updateReaction, removeReaction } from '../../../redux/thunks/feed'
+import { getPostReactions } from '../../../../../redux/selectors/feed'
+import { getUserId } from '../../../../user/store/userSelectors'
+import { createReaction, updateReaction, removeReaction } from '../../../../../redux/thunks/feed'
 import './style.scss'
 
 ReactionButton.propTypes = {
