+++ src/features/activity-feed/reaction/components/ReactionButton/style.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/settings/breakpoints' as b;
-@use '../../../styles/settings/sizes' as s;
+@use '../../../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/sizes' as s;
 
 .c-reaction-selector{
